+++ lib/transpec/util.rb
@@ -29,6 +29,7 @@ module Transpec
         namespace_node, name = *const_node
         const_names << name
         break unless namespace_node
+        break unless namespace_node.is_a?(Parser::AST::Node)
         break if namespace_node.type == :cbase
         const_node = namespace_node
       end