+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 * Continue processing files even if a file has invalid syntax
+* Fix a crash on source `variable::Const`
 
 ## v0.1.1