+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Master
 
+## v0.1.2
+
 * Continue processing files even if a file has invalid syntax
 * Fix a crash on source `variable::Const`
 * Fix generating invalid code with here document followed by method